requirements.txt
streamlit>=1.32.0
together>=0.2.11
pillow>=10.0.0 